requirements.txt
PyMuPDF==1.22.4
yara-python==4.3.1

